Moved part of the request police code, to the session module. Not working now, but need to do this, because I need to rebase on labs commit.
